backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.4 
celery==5.3.1
redis==4.5.5
django-redis==5.2.0
reportlab==4.0.4
gunicorn==20.1.0